Remove wasm-bindgen-backend
This internal crate doesn't have any external dependents, and I noticed that the "common functionality" is actually very neatly split between APIs only used by macro-support and APIs only used by wasm-bindgen-webidl.

Thus, in the spirit similar to https://github.com/wasm-bindgen/wasm-bindgen/pull/4608 I've moved this functionality to the crates that use it.

Since the utils from this crate are no longer all `pub`, this also allowed Rust to notice and mark some of them as entirely unused, so I removed those in the process.
