--- crates/backend/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-authors = ["The wasm-bindgen Developers"]
-description = """
-Backend code generation of the wasm-bindgen tool
-"""
-documentation = "https://docs.rs/wasm-bindgen-backend"
-edition = "2021"
-homepage = "https://wasm-bindgen.github.io/wasm-bindgen/"
-include = ["/LICENSE-*", "/src"]
-license = "MIT OR Apache-2.0"
-name = "wasm-bindgen-backend"
-repository = "https://github.com/wasm-bindgen/wasm-bindgen/tree/master/crates/backend"
-rust-version = "1.57"
-version = "0.2.104"
-
-[features]
-extra-traits = ["syn/extra-traits"]
-
-[dependencies]
-bumpalo = "3.0.0"
-log = "0.4"
-proc-macro2 = "1.0"
-quote = '1.0'
-syn = { version = '2.0', features = ['full'] }
-wasm-bindgen-shared = { path = "../shared", version = "=0.2.104" }
-
-[lints]
-workspace = true
