+++ crates/macro-support/Cargo.toml
@@ -18,10 +18,10 @@ extra-traits = ["syn/extra-traits"]
 strict-macro = []
 
 [dependencies]
+bumpalo = "3.0.0"
 proc-macro2 = "1.0"
 quote = '1.0'
 syn = { version = '2.0', features = ['visit', 'visit-mut', 'full'] }
-wasm-bindgen-backend = { path = "../backend", version = "=0.2.104" }
 wasm-bindgen-shared = { path = "../shared", version = "=0.2.104" }
 
 [lints]
