+++ crates/macro-support/src/codegen.rs
