+++ crates/macro-support/src/lib.rs
@@ -3,24 +3,25 @@
 
 #![doc(html_root_url = "https://docs.rs/wasm-bindgen-macro-support/0.2")]
 
-extern crate proc_macro2;
-extern crate quote;
 #[macro_use]
-extern crate syn;
-#[macro_use]
-extern crate wasm_bindgen_backend as backend;
-extern crate wasm_bindgen_shared as shared;
+mod error;
 
-pub use crate::parser::BindgenAttrs;
-use crate::parser::{ConvertToAst, MacroParse};
-use backend::{Diagnostic, TryToTokens};
+mod ast;
+mod codegen;
+mod encode;
+mod hash;
+mod parser;
+
+use codegen::TryToTokens;
+use error::Diagnostic;
+pub use parser::BindgenAttrs;
+use parser::{ConvertToAst, MacroParse};
 use proc_macro2::TokenStream;
 use quote::quote;
 use quote::ToTokens;
 use quote::TokenStreamExt;
 use syn::parse::{Parse, ParseStream, Result as SynResult};
-
-mod parser;
+use syn::Token;
 
 /// Takes the parsed input from a `#[wasm_bindgen]` macro and returns the generated bindings
 pub fn expand(attr: TokenStream, input: TokenStream) -> Result<TokenStream, Diagnostic> {
@@ -45,7 +46,7 @@ pub fn expand(attr: TokenStream, input: TokenStream) -> Result<TokenStream, Diag
 
     let opts = syn::parse2(attr)?;
     let mut tokens = proc_macro2::TokenStream::new();
-    let mut program = backend::ast::Program::default();
+    let mut program = ast::Program::default();
     item.macro_parse(&mut program, (Some(opts), &mut tokens))?;
     program.try_to_tokens(&mut tokens)?;
 
@@ -78,7 +79,7 @@ pub fn expand_class_marker(
     let mut item = syn::parse2::<syn::ImplItemFn>(input)?;
     let opts: ClassMarker = syn::parse2(attr)?;
 
-    let mut program = backend::ast::Program::default();
+    let mut program = ast::Program::default();
     item.macro_parse(&mut program, &opts)?;
 
     // This is where things are slightly different, we are being expanded in the
@@ -191,7 +192,7 @@ pub fn expand_struct_marker(item: TokenStream) -> Result<TokenStream, Diagnostic
 
     let mut s: syn::ItemStruct = syn::parse2(item)?;
 
-    let mut program = backend::ast::Program::default();
+    let mut program = ast::Program::default();
     program.structs.push((&mut s).convert(&program)?);
 
     let mut tokens = proc_macro2::TokenStream::new();
